Implement a replacement for reflect.DeepEqual for our tests.
Relying on reflect.DeepEqual has a couple big problems:
   - There is no way to make it compare maps where the keys are pointers
     to equivalent values, since it just does pointer comparisons.
   - There is no way to override how it does equality comparison for
     custom types, which means that for types that have pointer members
     again we are screwed.

Our own implementation of DeepEqual is simpler and does not rely on
reflection, which makes the code a bit more verbose, but prevents any
surprising behaviors.

Remove the helper function WrapComplexMapKey because for some reason
I don't understand, it was confusing the Go runtime, as sometimes some
maps and slices were given a runtime type of `*interface{}' instead of
`*[]interface{}' or `*map[interface{}]interface{}', which makes some
type assertions fail unexpectedly.  Might be a bug in the runtime or
some other limitation of the language that I don't understand.

Change-Id: I4ba2f18b88b336a47089841ca668ec794959f664