Make DeepEqual / Diff truly generic.
`DeepEqual()' and `Diff()' now can work with any type.

Our `DeepEqual()' remains faster than `reflect.DeepEqual()' because it
has a reflection-free fast path, but now it resorts to a reflection
based approach upon encountering an unknown type.

In addition, it is now possible to exclude certain attributes from the
consideration of `DeepEqual()' and `Diff()' by defining them with the
tag `deepequal:"ignore"`.

Change-Id: Ica32c93fd04a4f4ea4e36064d49a6bd29f4617ef