Allow `DeepEqual()' to switch back to the fast path.
Once `DeepEqual()' encounters a type it doesn't know how to handle, it
switches to using reflection to do the equality comparison.  The code
was never switching back to the faster non-generic code, even when it
was possible.  Now we can navigate our way into any kind of struct using
reflection but still benefit from the fast path once we're inside.

Change-Id: I4e1b12b9fa2265b79841273a667017e8c50385da