test: improve Diff()'s error messages.
By enforcing the convention that Diff() must be called with the expected
value first and the actual value second, we can have Diff() produce much
more helpful error messages.

I keep getting confused when seeing test failures that say "one thing
was X but the other was Y", because it's not immediately obvious which
one of X or Y was the value we wanted vs the actual one we got.

And to make things even more confusing, some tests were doing it one
way, others the other way.

Change-Id: I9b0cd48de0cb46307f6c50142e4f8df53f0a191f