Make all keys types.Key
Sorry for the humongous change, but changing either API forced me
to change all the other ones.
This is ground work to allow SetAttribute() to know if it needs to be
notifying or not: we will achieve that in a subsequent change. Having
types.Key instead of string in the collection APIs gives us more
flexibility on what to put in the key value (e.g., wrap it around a
new DontNotify type).

Change-Id: I0fd440f49550c68885e7653dcabe7a003457544d