Remove some unnecessary parentheses that don't help readability.
Courtesy of "gofmt -r '(a) -> a' -w", thanks Frank.
In cases involving binary operators, where gofmt removes the whitespace
around the operator to show it "binds more strongly" and readability is
still good, I opted to remove the parentheses as the spacing made it
obvious in which order operations would occur.

Change-Id: I6e355ff11e25af32692462e3bf43e49dc3defa70