key: Stringify() doesn't need to be exported.
Also it doesn't need to return an error, because in every single place
where we call it, we panic() when it returns an error.  Not exporting
it also means it won't receive random inputs, and key.New() already
validates the types of values it accepts as input, so panic()ing feels
like less of an issue now.