key: Add a sentinel to the composite keys to strengthen our hack.
Our hack is changing the typeAlg of all types that are exactly one word
long and are unhashable.  This is a bit risky in case someone creates
such a type and sticks it as a key in a map (the map would have to be
keyed by an interface type, though, otherwise the compiler would reject
the code).  In order to be a bit more defensive against such buggy code,
we pay the price of an extra word for hey composite key, that we then
check at runtime after casting from `unsafe.Pointer', to make sure that
we can flag such buggy code without crashing with an undefined behavior.

Gosh I hope nobody from the Go team will find our hacks.  They'll
probably create a license to write Go code and make sure we never
get one.

Change-Id: If751a272f99d25d9f8b4add154bd7e5b9d6d17d0