ocredis: Add support for authentication with clusters
Without using redisc.RetryConn, whenever we tried to write to a
different node in the cluster besides the node that received the
request, redis would return a MOVED error, and we would never store
the value into redis. However, just using redisc.RetryConn wasn't
enough either because redis would return authentication errors, even
though I had already called AUTH. My guess is that only one node in
the cluster would receive the AUTH. To fix this, I used a redis.Pool.
I don't fully understand how the redis.Pool works, but calling AUTH
from within the Pool fixed the authentication issue.

One limitation of this approach is that with the RetryConn, I'm forced
to use Do, which means we can no longer buffer and pipeline writes. I
will also see if there's a way to do this without using the RetryConn
so that we can continue to pipeline writes.

Change-Id: I9b00ff9673ced713e740b61cc5ae6e6920cb9384