openconfig/client: Panic if we don't get `sync_response: true`.
One of the symptoms of using a client built with the pre-release
openconfig.proto code against a newer server is that the sync response
comes back false, when it should always be true.  This has confused a
lot of people, so panic()'ing with an error message will help steer
people in the right direction faster.

Change-Id: Icb5bbcd9307278cf5f7dea7398d2feb4834bfb9e