key: Add stringify'd key to stringifying map[string]interface{} case
It's possible (in Turbines and other clients) for two different Keys
to be map[string]interface{}s with different keys sorted in the same
order with the same values.  With the current implementation, these
two different Keys will stringify to the same thing, and when we
JSONMarshal a notification, they will conflict and one will be
overridden.

Change-Id: I54e2db72bce5dbbcbaf24883e8638ab60ab47fea