monotime: add Since function
This is a convenience function for computing the elapsed time since a given
timestamp, as it's a little cumbersome to type

`time.Duration(monotime.Now() - start)`

over and over. Adds an example to show how it should be used.

This adds a dependency on the time package, which might not be desirable.