monotime: add Since function
This is a convenience function for computing the elapsed time since a given
timestamp, as it's a little cumbersome to type

`time.Duration(monotime.Now() - start)`

over and over. Adds an example to show how it should be used.

This adds a dependency on the time package, which might not be
desirable.

Patch contributed by @kevinburke via
https://github.com/aristanetworks/goarista/pull/14

Change-Id: Ie83a8746524be7ad384bcbc05dc8a0460153f412