kafka/producer: fix race condition
Method Run() used to be always called in a go routine, which would
result in wg.Add() also be called in a go routine. This is racy since
Stop() method uses wg.Wait() which could be called right after go
Run() which could be scheduled to run later.

To fix this, wg.Add() should always be called before wg.Wait(),
so the new method Start() is introduced that is non-blocking and assumed
to be always called before Stop() (which is usually the case).

Change-Id: Ib46dd704763049f444325b3983619e8df5dd1163