encoder: introduce kafka.MessageEncoder interface
benefits of adding this interface:
1) Let the encoder handles all metadata related operations instead of
inside the producer, e.g. HandleSuccesses and HandleErrors.
Hence, in the Encode() function, encoder can define what types of Metadata
that it wants to include in the sarama.ProducerMessage, and defines its
own way to deal with the metadata in the HandleSuccesses and
HandleFailures methods.
2) Let the encoder deal with monitoring intead of inside producer,
which lets producer serve only for producer related operations.
3) Each different typs of encoder can define its own NewEncoder()
based on what parameters it needs.

Change-Id: I7dd676a2dfb8e72afe503f1d8c3fc073f17e75c8