kafka/producer: Fix race condition in producer's Stop function
When we were calling the Stop function of the producer we were closing
it's done channel and then the kafkaProducer. The problem is,
it can happen that we will be in the function produceNotifications
when we will call stop, and since the select statement perform randomly
when multiple go channel are available to read from, it can happen that
we try to send a message to the kafkaProducer that will be already closed.

The fix here close the done channel, then wait for all the go routine to be
stopped and then close the kafkaProducer.

Change-Id: I9f1bd714a2a3c5e4f18e24546e54a53877e81c71