kafka/producer: Fix flakey test
The test calls Write on the producer in two different goroutines, but
no one was consuming that write. The test would only pass if the main
test goroutine won the race to be first to write. I added a consumer
to handle that.

In addition the test was spawning a goroutine that was never
stopped. I added cleanup.

Change-Id: If7d2cb3d3eeae15089904949af0dacc630e2f2d4