gnmi: pull in gnmi path parsing code from gopenconfig
The latest gnmi spec/proto (0.4) deprecates Path.Element, which
was a []string, and instead we'll now use a slice of PathElem,
which has a Name and optional map of keys. So basically it's now
up to the gnmi client to build the real gnmi path, rather than
using an intermediate string form over the wire.

This code was moved verbatim from gopenconfig/gnmi/server, and
then simplified a bit. We'll make use of it in the next change.

Change-Id: I661847b82885ce73c45b742037f8647e1e2ebcea