key: Wrap strings more efficiently.
keyImpl (sorry for the awful Java-style name) is a struct that contains
an interface{}, which is 2 pointers.  It itself implements the Key
interface, so is returned as an interface (again, 2 pointers).  This
means we have a double indirection going on there.

We wrap a lot of strings with key.New() and the second indirection costs
16 bytes per instance (8 for 32-bit binaries).  The good news is the
second indirection can be eliminated by having type-specific Key
implementations.

This change introduces a type-specific Key implementation for string
keys.  Technically we could (and probably should) do this for all the
built-in types supported by Key, but for now I'm just starting with one
of the most common ones.

I've been meaning to do this for a while but it was a bit harder to do
back when the Key interface was bloated with methods more complicated
to implement, which we thankfully removed in commit c8d7f36e.

Change-Id: I4d0caed098266d535aafe16baa7e6bf217f6c024