cmd/importsort: Determine first party code from VCS root directory
importsort required passing an ugly flag to tell it which is your
first party code. This change makes that unnecessary by looking for a
VCS root directory, eg. a .git directory, to determine which is first
party code. This automatic analysis is only done if the sections flag
is not passed.

Change-Id: If5fdc742e01b46cfba7d603d8752a096f7f33e18