cmd/importsort: Stop requiring import prefixes to end in "/"
The original implementation simply searched for a prefix of import
paths to establish section. Thus, if you wanted "foo/pkg" to sort in a
different section than "foobar/otherpkg" you needed to specify a
section prefix of "foo/". The problem with this is it doesn't match
the package "foo", which you probably wanted.

This change expects allows you to specify "foo" as the section prefix
and treats that as a whole package name.

Change-Id: I7521bffe5de278c69cac50682a9d3535be36089c