cmd/importsort: Check VCS root per file
Previously, the VCS root was determined once for the first file that
was passed in and assumed that all other files use the same VCS
root. This change implements finding the VCS root per-file, though the
result is cached per directory. This has not had a noticeable effect
on runtime for importsort when running across a single repository with
1542 go files and 499 directories, which was about 230ms on a 2014
MacBook Pro with SSD.

Change-Id: I1cf18d2a7679e9d2d1afa79d328909b50517af5f