path: Accept relative path strings with FromString method
This change removes the panic from FromString which was
preventing strings not leading with '/' from being used
as input. Now, any leading '/' in an input string will be
dropped before the string is split and constructed into a
Path. Note that strings not leading with '/' will not be
reconstructible with the String method.

Change-Id: If465e0b9be5b4d172ff5790e690ba7d4c5459fea