sizeof: Add DeepSizeoff function.
This function takes as a parameter a pointer to any type of data
structure and will analyse the total memory footprint of it.

To achieve this, this function will follow all structures, typed
pointers, channels and maps inside the given data structure and
analyse all types under them. For maps, the function will retrieve the
size of the struct hmap defined in file /go/src/runtime/hashmap.go
plus the size of the memory covered by the map (i.e. the size of
current buckets in use plus the size of the olds buckets if they
exist), it will then iterate the map and get the memory footprint of
each key/value association. For channels, the function will return the
size of struct hchan defined in file /go/src/runtime/chan.go plus the
size of the buffer mapped under the attribute buf in struct hchan
(computed by doing dataqsize * elemsize). It will then iterate on all
elements available inside the buffer and get their memory footprint.

In the end, this function return a map[string]uintptr where keys are
all types discovered and the associated value is the cumulative size
taken by those types discovered during the analysis of the given data
structure.

Change-Id: Ia0e17cc3a62d47059151fc4e0bef0cf98725137f