sizeof: Don't import glog.
First of all because we never use upstream glog in our code, we always
use our own fork.  Second, because a general purpose library shouldn't
import any logging library.  Third, because this error case is really
not a legitimate reason to log something, we really need to return an
error here.

I don't think there is a good reason to require a pointer in argument
here, we should be able to accept anything.  Something to improve later
on, I guess...

Change-Id: I129abfaef0cd50646ec67fbe56cff037b13913cb