path: Redefine Wildcard as a unique key.Key
With the previous implementation, paths containing a
key.New("*") would be interpreted by a Map as having
a wildcard in them. However, this can result in paths
unintentionally containing wildcards in them.

This change redefines a wildcard as an unexported struct
implementing key.Key. As a result, paths instantiated
outside this package must explicitly contain the Wildcard
variable in them to be interpreted as having a wildcard
in a Map.

Change-Id: Id9bc555b4653319682a8ad420d6209a51d3f02ad