path: Return (v, ok) with Map.Get instead of just v
Map currently allows us to set a path to nil, meaning
Map.Get will return nil in both the case that a path
is mapped to nil and in the case that a it just does
not exist. This change updates Map.Get to return both
a value and bool that indicates whether or not a path
exists in the map. For example:

a := New("foo", "bar")
b := New("bar", "baz")

m.Set(a, nil)

v, ok := m.Get(a) // v == nil, ok == true
v, ok = m.Get(b)  // v == nil, ok == false

Change-Id: Iafabd6ec8776b531c7747487ecc2f1e3984aa6f8