path: Update Equal and HasPrefix
Change 1:

Convert Path.Equal to Equal. There has not been
a real use case for having Path implement the
key.Comparable interface. We can also avoid the
type assertion by making this change.

Change 2:

Convert Path.HasPrefix to HasPrefix. This should
have been done from the start as Path.HasPrefix
is not implementing any interface and it is just
as easy to understand HasPrefix with respect to
the name of this package.

Change 3:

Add some test cases pertaining to nil paths for
Equal and HasPrefix. That is, test that nil and
empty paths are treated the same:

Equal(nil, Path{})
Equal(Path{}, nil)
HasPrefix(Path{}, nil)
HasPrefix(nil, Path{})

Change-Id: Ia7c22e8e1171e9073d28ece945d222960ecd3544