path: Add Map.VisitPrefixed
Define a new function on Map called VisitPrefixed that calls a
handler on all values registered with a path that is prefixed
by a given path. This method can be used to visit all values in
the Map simply by passing it the empty path (which is a prefix
of all paths).

Change-Id: I4729e877ca89bb14ceb1ad8bd81238a87ef0abcf