key: Do not escape string key when using the String() method
String can either represent a valid string or a byte array in Aeris.

1/ We change the String() of string key. It now returns the exact string
without any escaping. If the string is a binary blob it's the
responsibility of the caller to deal with it.

2/ We change the way we escape the strings:
If a string is a valid UTF-8 string we don't change anything and use
this string.
If a string is not a valid UTF-8 string we base64 the string and return
this base64 encoded representation of the binary string.

Change-Id: I71da6eb017643f51818d57e91959cea8eda5e1cf