path: Improve Wildcard object and type
The Wildcard (fake) constant should not implement key.Key interface but
value.Value interface.
Then it can be used to build a Key.
So key.New(Wildcard) should be allowed and it should also
be stringified using StringifyInterface.
In order to allow this, the "wildcard" type implements value.Value
interface.
The reason to have WildcardType exported is that some packages might
need to type cast keys and detect WildcardType keys, so it seems
important to let other packages detect this type.

Change-Id: Iefb04b0155d49085309ef7d3ec61d118ad828764