key: Implement key.Pointer
In order to make this implementation usable as a key,
we have to add another hack to the key package. This
hack special cases key.Pointer in key.New. For the
pointers being used as top level keys, they are converted
to a special type 'pointerKey' type that is a composite key.
When pointerKey.Key is called, a key.Pointer is constructed
and returned. For the pointers being used as values inside
composite keys, they are converted to slices and then hashed.

Change-Id: I9ca957376ff0b6b20316953c831945e5223fddc0