ocprometheus: Add string handling
OCPrometheus ignores non float64 values when parsing updates.
This commit changes the parseValue function so that it returns an
interface containing either a string or a float64.
Comments have been added to clarify what the function returns.

If a metric has a value label it is appended to its labels and the value
changed whenever it receives a string update.

The update function now handles string values from parseValue.
If parseValue returns a string, the string is stored in the value label
of the metric. Prometheus cannot display strings otherwise and drops
them silently if they are stored in the value part of the metric. String
metrics show a default value that is configured per metric.

The config now takes in a label to be used to store string values and
a default value to display in the metric. If it has a value label then
the metric is marked as being a string metric.

getDescAndLabels changed to return a struct of all the values that are
needed for updating or creating a metric. Now named getMetricValues

Sampleconfig file updated to include example of label value change and
default value.

Fixed typos in a log message and a comment.

Edited tests to reflect changes.
- Collector_test.go wasn't saving changes to labels when it was creating
  each set of expected metrics. Now uses the previous metrics and
  updates those instead of creating new ones.

- Config_test.go now has label values and default values.

Change-Id: If209bf6f369534b501fb85d3bf7a0fcad9e3116d