all: Replace bytes.Buffer with strings.Builder
strings.Builder was introduced in go1.10. It is equivalent to a
bytes.Buffer, except that String() is free (it doesn't do an
allocation), and you can't read from it other than with String().

I noticed in one CPU graph of TerminAttr usage that path.String() was
taking a non-trivial amount of CPU and a simple way to decrease that
amount was to use a strings.Builder. I then went ahead and replaced
all bytes.Buffers with strings.Builders.

Change-Id: Ibc1561d0660cdf10ccd5d9996658a27c960fb501