cmd/octsdb: Include metric key in metric name
This fixes a confusing behavior. If there are no non-named
subexpressions in a metric then the metric has no name and matches are
ignored. For example, the following would never match anything

"/Sysdb/environment/temperature/status/tempSensor/(?P<sensor>.+)/temperature"

because Match would return "" for the metric name.

In addition it seems odd to me that the metric key from the JSON
wouldn't be included in the metric name.

Change-Id: I92def679e53ab86aeb4ec95163da17381b6b48ea