cmd/ocprometheus: Deal with coalesced deletes
Sometimes TemrinAttr can coalesce collection deletes such that only the
delete all for the collection is sent.
That means we can't rely on deleting the exact metric path, but instead
we need to delete all the metrics that are in the collection/subtree.
For now I implemented a simple linear scan, if the performance is a
problem we should move to a prefix tree approach.

Change-Id: Ie0a8d45f2925b9abff370d10ca2c8f7226188d74