path: add path.HasPrefixString()
HasPrefixString returns whether path b is a prefix of path a.
It checks that b is at most the length of path a and
that every element of a matches corresponding element of b exactly
except the last element.
The last element of a is considered a match if it matches exactly
or if it is a string and the last element of a is a subsring
of corresponding element of b

Change-Id: I526155d9074fa2b1f374f4ab5e688221b8d0f4c5