path: Add path.MatchPrefixString()
MatchPrefixString returns whether path b is a prefix of path a
where path a may contain wildcards.
It checks that b is at most the length of path a and
that every element of a matches corresponding element of b exactly
or a wildcard except the last element.
The last element of a is considered a match if it matches exactly
or it is a wildcard or if it is a string and the last element of a
is a substring of corresponding element of b

Change-Id: Ief8a3d79c69d95121fa072d7d89591167888660c