key: add new Map type
A Map represents an arbitrary map with keys of arbitrary type. If the
key of an entry in a pending Set() is not natively hashable (as is the
case when keys themselves are either maps or slices), the key must
implement the Hashable interface before it can be added to a Map.
Also add a Set() method to add entries.

Change-Id: Ia44b68fe35c1035964b1822c1977135e5ec86e08