Actually fix go1.12
pinky promise this is it.

Here's the problem;
go1.13 has GOSUMDB and GOPROXY set up. Using go1.13, tjfoc/gmsm@v1.3.0
was downloaded from the proxy. go1.13 verified the thing it downloaded
matched the checksum from GOSUMDB, wrote the checksum in sum.go, and
went about its way.

Now, telling go1.12 to download the modules fails on the checksum of
the above package. The issue being that for some reason, downloading
gmsm from github, its checksum is different from what GOSUMDB has. And
why does go1.12 not use proxy.golang.org? Because GOPROXY is not set
in that version of go. Setting GOPROXY makes go1.12 download the right
version of gmsm and this all goes away.

Change-Id: I7f7068c4ce2b4955383700471411cd061c0712db