gnmi: Add support for target field
I've added a command-line option for `target`, and passed the specified
target on to Get, Set, and Subscribe. This is very similar to `origin`
except that according to the spec (2.2.2.1), the target "MUST only ever
be present on prefix paths in the corresponding request and response
messages," so a prefix is created whenever a target is specified.

Change-Id: I78f2fd2bfa586e1a924ef2c111804cc43dc79af8