gnmi: Allow empty strings for keys and values in element keys
gNMI paths are composed of path elements, each made up of a name and optionally
a set of keys represented as a map. The requirement from the spec is that these
are UTF-8 encoded strings. The empty string is still a string, so allow empty
strings as keys and values.

Change-Id: I265401709bdf82faa7c8d575a932a6ace3d93dca