cmd/ocprometheus: use path.Join to concatenate prefix and update path.
This was causing ocprometheus to error out because it seems that
notifications coming from a gnmi subscription to OpenConfig
can have their prefix set to nil or not depending on some
circumstances.
i.e we can receive two notifications one after the other (or not)
that would look like this:
n1 = Prefix: foo; Updates: {{Path: {bar}, Value: 1}}
n2 = Prefix: nil; Updates: {{Path: {foo, bar}, Value: 2}}

When nil, ocprometheus was adding a new key for a metric that would
already exist in its metrics map if it received a notification for the
same metric where the prefix path was not nil. That is because the
key for this map is a string that is the result of the concatenation
of <notif prefix> + <update path>. Since we were returning "/" for
a nil prefix, the key was looking like "//foo/bar" whereas it should
be "/foo/bar". Then the prometheus library would complain about having
handled the same metric with the same labels twice.

Change-Id: Ide092b8628943996332716e7ef5dc644fd2b9973