gnmi, dscp: Drop dependency on github.com/aristanetworks/glog
The gnmi package is often used in combination with the
github.com/openconfig/... packages which depend on
github.com/golang/glog. These two glog packages cannot be in the same
binary because they both try to register the same flags.

This change adds a test to ensure gnmi doesn't depend on
aristanetworks/glog and removes one of those dependencies in the dscp
package.

Change-Id: Iff4259b667eceea33671b7e16ce1fdef8a926141