gnmireverse/client: Add Subscribe mode for gNMI Get dial-out
Currently, the gNMIReverse client issues a gNMI Get to the target at
every sample interval and sends the GetResponse to the gNMIReverse
collector server.

Add the option to perform a periodic Subscribe sync instead of a Get,
which gathers the notifications from the sync and bundles them into
one GetResponse. This gathers individual leaf updates (instead of a
subtree with Get) and preserves the timestamps for each leaf.
- For OpenConfig paths, keep a Subscribe POLL stream to perform a sync
  at each sample interval.
- For EOS native paths, because of EOS backwards compatibility, perform
  a Subscribe ONCE if it is supported, otherwise, perform a Subscribe
  STREAM and close the stream after the sync is complete.

  -get_mode string
      Operation mode to gather notifications for the GetResponse message.
        get        Gather notifications using Get.
        subscribe  Gather notifications using Subscribe.
                   Notifications from the Subscribe sync are bundled into one GetResponse.
                   With Subscribe, individual leaf updates are gathered (instead of
                   a subtree with Get) and timestamps for each leaf are preserved.
       (default "get")

Change-Id: Ie10f0ab05725d319a062d705e1c0f9d7c8492f5f