netns: Panic in Do if fails to return to original namespace
If netns.Do fails to return to the original namespace the process is
in a very weird state, with one thread controlled by the Go runtime
stuck in a different namespace from all other threads. Goroutines
jumping between threads will sometimes be in one namespace sometimes
another. It is best to panic here and kill the process.

Error wrapping is also added in this review so that users of netns.Do
can detect the error that was hit.

Change-Id: I6a1adc70ad759bb151ac14a49b19d643d42c4519