modroot: Add it
modroot.Path() returns the full path to the module root. This is meant
as a way for tests to get absolute paths to test fixtures or interesting
files. The problem addressed usually arises when a test helper reads
a file, and said helper is called from different test packages. The path
to the file will then depend on the package being tested, not the
helper. Each package will have to pass the helper a different
relative path or the helper will have to resort in gruesome solutions
like runtime.Caller or even worse, expecting the whole repo to be under
GOPATH and GOPATH to be set.

Change-Id: I1aa5bd6b12762ae77ea24d7561d78a0c94b5b263