cmd/ocprometheus: create prometheus metric on the fly
previously we created the prometheus description once and then stored
it. There is a new feature being added which allows us to dynamically
extract tags from description leafs on the fly, this means we need to be
able to update the prometheus description on the fly. As such I have
added a struct which contains the information required to do this.
A follow up commit will contain the usage of this change.

Change-Id: I448c4ecd943a993130001be5ef415713b3e13e7e