everything: bump to go1.20 and update dependencies
Update go.mod version, pipelinevars and update_deps.sh.
Run update_deps.sh to pull go1.20 modules.

Pulling the latest version of github.com/prometheus/client_golang
was causing a test to fail because test.DeepEqual was not able to
handle comparison between function types. (these are incomparable
types in the go compiler - it doesn't generate the necessary
equality functions to be able to do '==' comparisons).

To fix this, make test.DeepEqual return true when comparing functions
if the functions are both nil, or if the functions have the same
name.

Change-Id: Iaf3a8cd7f5acd780a1cc07dfd92cbeed057bc4dd