cmd/octsdb: Don't subscribe to / by default
octsdb will subscribe to / (root) by default. I believe this was a
mistake because there is even a comment about filtering out this
default path, but the code stopped working after some refactors. This
subscribe to / causes a lot of extra load for the gNMI server and
octsdb. The only way to undo that is to pass a -paths flag, but most
users won't be doing this since the subscription paths are set by the
config file.

Change-Id: I9b8ccf3b5fa5c14ccc6d3f1937c54a6272c7f608