gnmi: add optional flags to specify min/max TLS versions
These flags are optional. Without specifying any of them,
fallback to the default behaviour enforced by grpc-go.

The supported min/max versions are dynamically generated
based on cipher suites supported by the go standard library
crypto/tls package (which is used by grpc-go).

This commit also adds a test to ensure that the regex used to
dynamically generate the names of these cipher suites
(for use in flags) is sufficient (not under-specific
or failing to match)

running the help string for the gNMI client now prints
  -tls-max-version string
        Set maximum TLS version for connection (1.0, 1.1, 1.2, 1.3)
  -tls-min-version string
        Set minimum TLS version for connection (1.0, 1.1, 1.2, 1.3)

Change-Id: If741f3284d3593ff212497fc57e902250d6549f8