gnmi/client: Add a manual version
Due to BUG905740, the output of `gnmi -version` is blank. A hard-coded
version number will fix this. The version number is the date of the last
change.

As a small sanity check, I built the binary and ran the cmd manually:

`~/tmp/go/src/goarista/cmd/gnmi @adityasingh> go build
~/tmp/go/src/goarista/cmd/gnmi @adityasingh> ./gnmi -version
2024.10.20 d437ec8d8de432bfd0c297f104db4489a49d23af_go1.22.6`

Fixes: BUG905740
Change-Id: I5cfcccc9e4fb4833ebc1fb0fcc241d6eb65f8c0a