gnmireverse/client: Support dial to Unix domain socket gNMI server
The gNMIReverse client uses the same ParseAddress/Dialer functions
used by the gNMI client to dial to the target gNMI server. This allows
for connections to Unix domain socket addresses.

The default value for "-target_addr" is changed to
"unix:///var/run/gnmiServer.sock". As a result, the following flags
become redundant as the Unix domain socket gNMI server is inherently
secure (as it is on the same device) and does not require
authentication:

  -username
  -password
  -credentials_file
  -target_addr
  -target_cafile
  -target_certfile
  -target_keyfile
  -target_tls_insecure

"-collector_addr" also uses the same functions for completeness.

Fixes: BUG1079430
Change-Id: Ib9c4c993356235278e14da3d3720b95199024856