sizeof: Drop support for maps in go1.24
Go's map implementation was rewritten in go1.24, this caused sizeofmap
to panic when it couldn't find the expected type. Skip over this case.

To truly support go1.24 maps would require a rewrite of sizeofmap that
relies on deep internals of the map implementation. Personally, I
would rather deprecate this whole package as it doesn't have much use
and its results cannot always be trusted.

Change-Id: Iaac0637a202a55fba7590eb5e39f63ce810d3a93